Move def interp()to dpnp_iface_mathematical
