+++ dpnp/dpnp_iface_statistics.py
@@ -49,7 +49,6 @@ import dpnp
 
 # pylint: disable=no-name-in-module
 import dpnp.backend.extensions.statistics._statistics_impl as statistics_ext
-import dpnp.backend.extensions.ufunc._ufunc_impl as ufi
 from dpnp.dpnp_utils.dpnp_utils_common import (
     result_type_for_device,
     to_supported_dtypes,
@@ -67,7 +66,6 @@ __all__ = [
     "corrcoef",
     "correlate",
     "cov",
-    "interp",
     "max",
     "mean",
     "median",
@@ -1051,181 +1049,6 @@ def cov(
         fweights=fweights,
         aweights=aweights,
     )
-
-
-def interp(x, xp, fp, left=None, right=None, period=None):
-    """
-    One-dimensional linear interpolation for monotonically increasing sample points.
-
-    Returns the one-dimensional piecewise linear interpolant to a function
-    with given discrete data points (`xp`, `fp`), evaluated at `x`.
-
-    Parameters
-    ----------
-    x : array_like
-        The x-coordinates at which to evaluate the interpolated values.
-
-    xp : 1-D sequence of floats
-        The x-coordinates of the data points, must be increasing if argument
-        `period` is not specified. Otherwise, `xp` is internally sorted after
-        normalizing the periodic boundaries with ``xp = xp % period``.
-
-    fp : 1-D sequence of float or complex
-        The y-coordinates of the data points, same length as `xp`.
-
-    left : optional float or complex corresponding to fp
-        Value to return for `x < xp[0]`, default is `fp[0]`.
-
-    right : optional float or complex corresponding to fp
-        Value to return for `x > xp[-1]`, default is `fp[-1]`.
-
-    period : None or float, optional
-        A period for the x-coordinates. This parameter allows the proper
-        interpolation of angular x-coordinates. Parameters `left` and `right`
-        are ignored if `period` is specified.
-
-    Returns
-    -------
-    y : float or complex (corresponding to fp) or ndarray
-        The interpolated values, same shape as `x`.
-
-    Raises
-    ------
-    ValueError
-        If `xp` and `fp` have different length
-        If `xp` or `fp` are not 1-D sequences
-        If `period == 0`
-
-    See Also
-    --------
-    scipy.interpolate
-
-    Warnings
-    --------
-    The x-coordinate sequence is expected to be increasing, but this is not
-    explicitly enforced.  However, if the sequence `xp` is non-increasing,
-    interpolation results are meaningless.
-
-    Note that, since NaN is unsortable, `xp` also cannot contain NaNs.
-
-    A simple check for `xp` being strictly increasing is::
-
-        np.all(np.diff(xp) > 0)
-
-    Examples
-    --------
-    >>> import dpnp as np
-    >>> xp = np.array([1, 2, 3])
-    >>> fp = np.array([3 ,2 ,0])
-    >>> x = np.array([2.5])
-    >>> np.interp(2.5, xp, fp)
-    1.0
-    >>> np.interp([0, 1, 1.5, 2.72, 3.14], xp, fp)
-    array([3.  , 3.  , 2.5 , 0.56, 0.  ])
-    >>> UNDEF = -99.0
-    >>> np.interp(3.14, xp, fp, right=UNDEF)
-    -99.0
-
-    Plot an interpolant to the sine function:
-
-    >>> x = np.linspace(0, 2*np.pi, 10)
-    >>> y = np.sin(x)
-    >>> xvals = np.linspace(0, 2*np.pi, 50)
-    >>> yinterp = np.interp(xvals, x, y)
-    >>> import matplotlib.pyplot as plt
-    >>> plt.plot(x, y, 'o')
-    [<matplotlib.lines.Line2D object at 0x...>]
-    >>> plt.plot(xvals, yinterp, '-x')
-    [<matplotlib.lines.Line2D object at 0x...>]
-    >>> plt.show()
-
-    Interpolation with periodic x-coordinates:
-
-    >>> x = [-180, -170, -185, 185, -10, -5, 0, 365]
-    >>> xp = [190, -190, 350, -350]
-    >>> fp = [5, 10, 3, 4]
-    >>> np.interp(x, xp, fp, period=360)
-    array([7.5 , 5.  , 8.75, 6.25, 3.  , 3.25, 3.5 , 3.75])
-
-    Complex interpolation:
-
-    >>> x = [1.5, 4.0]
-    >>> xp = [2,3,5]
-    >>> fp = [1.0j, 0, 2+3j]
-    >>> np.interp(x, xp, fp)
-    array([0.+1.j , 1.+1.5j])
-
-    """
-
-    dpnp.check_supported_arrays_type(x, xp, fp)
-
-    if xp.ndim != 1 or fp.ndim != 1:
-        raise ValueError("xp and fp must be 1D arrays")
-    if xp.size != fp.size:
-        raise ValueError("fp and xp are not of the same length")
-    if xp.size == 0:
-        raise ValueError("array of sample points is empty")
-    if not x.flags.c_contiguous:
-        raise NotImplementedError(
-            "Non-C-contiguous x is currently not supported"
-        )
-    x_dtype = dpnp.common_type(x, xp)
-    if not dpnp.can_cast(x_dtype, dpnp.default_float_type()):
-        raise TypeError(
-            "Cannot cast array data from"
-            f" {x_dtype} to {dpnp.default_float_type()} "
-            "according to the rule 'safe'"
-        )
-
-    if period is not None:
-        # The handling of "period" below is modified from NumPy's
-
-        if period == 0:
-            raise ValueError("period must be a non-zero value")
-        period = dpnp.abs(period)
-        left = None
-        right = None
-
-        x = x.astype(dpnp.default_float_type())
-        xp = xp.astype(dpnp.default_float_type())
-
-        # normalizing periodic boundaries
-        x %= period
-        xp %= period
-        asort_xp = dpnp.argsort(xp)
-        xp = xp[asort_xp]
-        fp = fp[asort_xp]
-        xp = dpnp.concatenate((xp[-1:] - period, xp, xp[0:1] + period))
-        fp = dpnp.concatenate((fp[-1:], fp, fp[0:1]))
-        assert xp.flags.c_contiguous
-        assert fp.flags.c_contiguous
-
-    # NumPy always returns float64 or complex128, so we upcast all values
-    # on the fly in the kernel
-    out_dtype = fp.dtype
-    output = dpnp.empty(x.shape, dtype=out_dtype)
-    idx = dpnp.searchsorted(xp, x, side="right")
-    left = fp[0] if left is None else dpnp.array(left, fp.dtype)
-    right = fp[-1] if right is None else dpnp.array(right, fp.dtype)
-
-    idx = dpnp.array(idx, dtype="uint64")
-
-    queue = x.sycl_queue
-    _manager = dpu.SequentialOrderManager[queue]
-    mem_ev, ht_ev = ufi._interpolate(
-        x.get_array(),
-        idx.get_array(),
-        xp.get_array(),
-        fp.get_array(),
-        # left,
-        # right,
-        output.get_array(),
-        queue,
-        depends=_manager.submitted_events,
-    )
-    _manager.add_event_pair(mem_ev, ht_ev)
-
-    return output
 
 
 def max(a, axis=None, out=None, keepdims=False, initial=None, where=True):
